Create struct-array.c
C program that determines the average of quiz scores of all students in a course. The input data of the course will be stored in an input file, and a record of each student will include the following
The student’s name
The student's ID
Five quiz scores
An average of the students' quiz scores
The file will have up to (but no more than) 30 records. The first line in the file will indicate the total number of records in the file. An example file is given below, but you should make your own test files, too.