Create struct-search.c
C program to find the customer’s information from a bank data. The input data will be stored in an input file, and a record of each customer will include the customer’s name, account number, and current balance. You must store each customer's record in a struct. Since there are several customers in a bank, you have to use an array of structs to keep all records. 

The file will have up to (but no more than) 30 records. The first line in the file will indicate the total number of records in the file. An example file is given below, but you should make your own test files, too.

Your program should:
Ask for an input filename and read the input data
Present the user with a record-finding function
Let the user enter a name
If the name is found once, present all of the customer information
If the name is found multiple times, present all of the customer information (for all customers with that name)
If the name is not found, present an error message to the user
After a search, allow the user to choose to continue (search again) or not (exit)