Create array-rotation.c
C program that rotates an array to the left or right depending on user input. You will read your inputs from a file. You must use dynamic arrays to solve this problem. Make sure you free the array at the end of the program. Your code will be tested with valgrind to make sure there are no memory leaks.
Left rotation:
A left rotation operation on an array shifts each element in the array one unit to the left. The first element in the array must get copied into the last index position in the array.
For example: 
[1, 2, 3, 4, 5] after one left rotation becomes [2, 3, 4, 5, 1]
[1, 2, 3, 4, 5] after two left rotations becomes [3, 4, 5, 1, 2]
Right rotation:
A right rotation operation on an array shifts each element in the array one unit to the right. The last element in the array must get copied into the first index position in the array.
For example: 
[1, 2, 3, 4, 5] after one right rotation becomes [5, 1, 2, 3, 4]
[1, 2, 3, 4, 5] after two right rotations becomes [4, 5, 1, 2, 3]
You will read the array elements from a file input.txt. The first line in the file contains two numbers and a letter:
<n> <r> <d>
n: number of elements in the array
r: number of rotations
d: direction of rotation - can be L or R
The next line will have n numbers separated by space where each of the numbers n1, n2, n3 etc. are elements of the array.
	<n1> <n2> <n3> …