Create dll-ordered-list-stretch.c
Write a program that defines a struct named OrderedList to store integers in an ascending order. The initial list of integers must be read from a file. You must use a doubly linked list to solve this problem.

Your program should do the following:
Read the first line in the file that indicates the number of nodes (integers) in the file.
Read the integers from the file and store each integer as a separate node. The integers in the file will be in ascending order. 
The list must have no duplicate integers.
You must have a struct called Node to store the value of the integer, a pointer to the next element in the list and a pointer to the previous element in the list.
The ordered list must have a head and a tail pointing to the head/tail of the list.
Each new integer read from the file should be added to the tail of the list.
Display the following menu to the user to make a choice:

Display all nodes
Display nth node
Append a new node - Head or Tail? (H/T)
Append a new node in the Nth position
Delete a node - Head or Tail? (H/T)
Delete Nth node
Find a node
Get number of nodes
Display even numbered nodes
Display odd numbered nodes
Reverse the list